requirements.txt
django-allauth==0.37.0
django-crispy-forms==1.7.2
django-compressor==2.2
django-extensions==2.1.2
django-tables2==2.0.3
django-webpack-loader
Django==2.1.5
Werkzeug
djangorestframework
mysqlclient==1.3.13
django-redis==4.5.0
django-debug-toolbar==2.1
celery==4.3.0
redis==3.2.1
